+++ src/main/java/com/app/bookstore/service/book/BookServiceImpl.java
@@ -1,4 +1,4 @@
-package com.app.bookstore.service.impl;
+package com.app.bookstore.service.book;
 
 import com.app.bookstore.dto.BookDto;
 import com.app.bookstore.dto.CreateBookRequestDto;
@@ -6,7 +6,7 @@ import com.app.bookstore.exception.EntityNotFoundException;
 import com.app.bookstore.mapper.BookMapper;
 import com.app.bookstore.model.Book;
 import com.app.bookstore.repository.book.BookRepository;
-import com.app.bookstore.service.BookService;
+import com.app.bookstore.service.book.BookService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
